crud
Crud de Users en funcionamiento

+++ backend/src/main/java/com/ecommerce/e_commerce/controller/UserController.java
@@ -1,30 +1,34 @@
 package com.ecommerce.e_commerce.controller;
 
 import com.ecommerce.e_commerce.exception.ResourceNotFoundException;
-import com.ecommerce.e_commerce.model.User;
+import com.ecommerce.e_commerce.model.Product;
+import com.ecommerce.e_commerce.model.Users;
 import com.ecommerce.e_commerce.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.security.SecurityProperties.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
-@RestController
-@RequestMapping("/users")
+@Controller
 public class UserController {
 
     @Autowired
     private UserService userService;
 
-    @GetMapping
-    public ResponseEntity<List<User>> getAllUsers() {
+    @GetMapping("/users")
+    public ResponseEntity<List<Users>> getAllUsers() {
         return ResponseEntity.ok(userService.getAllUsers());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+    public ResponseEntity<Users> getUserById(@PathVariable Long id) {
         try {
             return ResponseEntity.ok(userService.getUserById(id));
         } catch (ResourceNotFoundException e) {
@@ -32,13 +36,24 @@ public class UserController {
         }
     }
 
-    @PostMapping
-    public ResponseEntity<User> createUser(@Valid @RequestBody User user) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(userService.createUser(user));
+    @GetMapping({ "/createuser" })
+    public String Create(Model model) {
+        model.addAttribute("user", new Users());
+        return "crudusers";
     }
 
+    @PostMapping("/createuser")
+    public String save(@Valid Users user, BindingResult br, Model model) {
+        if (br.hasErrors()) {
+            return "crudusers";
+        }
+        userService.createUser(user);
+        return "redirect:/"; 
+    }
+
+
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @Valid @RequestBody User userDetails) {
+    public ResponseEntity<Users> updateUser(@PathVariable Long id, @Valid @RequestBody Users userDetails) {
         try {
             return ResponseEntity.ok(userService.updateUser(id, userDetails));
         } catch (ResourceNotFoundException e) {
@@ -46,13 +61,11 @@ public class UserController {
         }
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-        try {
-            userService.deleteUser(id);
-            return ResponseEntity.noContent().build();
-        } catch (ResourceNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
+    @GetMapping("/delete/{id}")
+    public String eliminar(@PathVariable Long id) throws ResourceNotFoundException {
+      if (id > 0) {
+        userService.deleteUser(id);
+      }
+      return "Usuario borrado con exito.";
     }
 }
